requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.15.2
    # via -r requirements.in
amqp==5.3.1
    # via kombu
billiard==4.2.1
    # via celery
celery[redis]==5.4.0
    # via
    #   -r requirements.in
    #   celery
click==8.2.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
greenlet==3.2.2
    # via sqlalchemy
kombu==5.5.3
    # via celery
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
prompt-toolkit==3.0.51
    # via click-repl
psycopg2-binary==2.9.10
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via celery
python-decouple==3.8
    # via -r requirements.in
redis==5.2.1
    # via
    #   -r requirements.in
    #   celery
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.40
    # via
    #   -r requirements.in
    #   alembic
typing-extensions==4.13.2
    # via
    #   alembic
    #   sqlalchemy
tzdata==2025.2
    # via
    #   celery
    #   kombu
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
